+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 - Fixes a bug that allows relationships to be created between parents and children.
 - Added support for perspectives on elements and relationships.
+- Added support for element border colours.
 
 ## 0.9.2 (16th September 2019)